Fix: source map for Emotion CSS
I was wrong about Emotion CSS "working out of the box" in the last
commit. Emotion always work even installed without the Babel plugin, but
what does not work is the source map, which requires the plugin.

This also means that Vite's React plugin has to be the Babel version,
not the other one with SWC.